PROYECTO FINALIZADO

+++ JAVA/VanguardBank2023/src/java/interfaces/CuentaCRUD.java
@@ -11,7 +11,7 @@ import model.Cuenta;
 public interface CuentaCRUD {
     public List listarCuentas();
     public boolean agregarCuenta(Cuenta nuevaCuenta);
-    public Cuenta buscarCuenta(long numCuenta);
+    public Cuenta buscarCuenta(int numCuenta);
     public boolean editarCuenta(Cuenta cuenta);
     public void eliminarCuenta(long numCuenta);
 }
